requirements.txt
boto3
Flask
requests
zappa
awscli
wordcloud
pytrends
Pillow
pyntone
python-dotenv